feat(fileGenerator): move PDF generation logic from pdfGenerator to fileGenerator and update import in PaymentStatusScreen
